Fix hot module replacement, add react hot loader
