+++ .babelrc
@@ -1,7 +1,7 @@
 {
   "presets": [
     [
-      "@babel/env",
+      "@babel/preset-env",
       {
         "targets": {
           "firefox": "60",
@@ -14,6 +14,7 @@
     "@babel/preset-react"
   ],
   "plugins": [
-    [ "@babel/plugin-proposal-class-properties" ]
+    "@babel/plugin-proposal-class-properties",
+    "react-hot-loader/babel"
   ]
 }
