+++ package.json
@@ -25,6 +25,7 @@
     "@babel/polyfill": "^7.0.0",
     "@material-ui/core": "^3.5.1",
     "auth0-lock": "^11.12.1",
+    "connected-react-router": "^6.0.0",
     "history": "^4.7.2",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
@@ -64,8 +65,9 @@
     "postcss-loader": "^3.0.0",
     "postcss-preset-env": "^6.4.0",
     "prop-types": "^15.6.2",
+    "react-hot-loader": "^4.3.12",
     "react-test-renderer": "^16.6.3",
-    "redux-devtools": "^3.4.2",
+    "redux-devtools-extension": "^2.13.7",
     "style-loader": "^0.23.1",
     "webpack": "^4.26.0",
     "webpack-cli": "^3.1.2",
