+++ src/rootReducer.js
@@ -1,6 +1,8 @@
 import { combineReducers } from 'redux';
+import { connectRouter } from 'connected-react-router';
 import auth from './components/Auth/reducer';
 
-export default combineReducers({
+export default history => combineReducers({
+  router: connectRouter(history),
   auth,
 });
